+++ libnopegl/src/backends/vk/hwmap_videotoolbox_darwin_vk.m
@@ -47,7 +47,7 @@
 
 struct format_desc {
     int layout;
-    int nb_planes;
+    size_t nb_planes;
     struct {
         int format;
     } planes[2];
@@ -123,7 +123,7 @@ static int vt_darwin_map_frame(struct hwmap *hwmap, struct nmd_frame *frame)
         return NGL_ERROR_GRAPHICS_GENERIC;
     }
 
-    for (int i = 0; i < vt->format_desc.nb_planes; i++) {
+    for (size_t i = 0; i < vt->format_desc.nb_planes; i++) {
         struct texture *plane = vt->planes[i];
         struct texture_vk *plane_vk = (struct texture_vk *)plane;
 
@@ -136,7 +136,7 @@ static int vt_darwin_map_frame(struct hwmap *hwmap, struct nmd_frame *frame)
         CVMetalTextureRef texture_ref = NULL;
         CVReturn status = CVMetalTextureCacheCreateTextureFromImage(NULL, vt->texture_cache, cvpixbuf, NULL, mtl_format, width, height, i, &texture_ref);
         if (status != kCVReturnSuccess) {
-            LOG(ERROR, "could not create texture from image on plane %d: %d", i, status);
+            LOG(ERROR, "could not create texture from image on plane %zd: %d", i, status);
             return NGL_ERROR_GRAPHICS_GENERIC;
         }
